Feature/binary analyzer changes (#4511)
* WIP

* layout of dev

* wip

* Removed versions

* working but not tested

* local

* Local, package failing because of none timestamp in PREV

* WIP

* search tests ok

* Fixing tests and small refactor

* renamed module

* review

* ISO in client, reviewed

* Fixing tests

* Fixing tests

* New tests about metadata being updated

* Review

* Time cleared on export/package

* Fix revisions list

* fixed test

* Unused import

* WIP

* Review

* Fixed test message assert

* Use env var instead of test client

* Cool testing framework and a lot of remove with revisions tests

* Remove tests revisions

* Drafting tests

* Test passing, a lot missing yet

* PENDING MANY TESTS

* some more tests

* Pending many tests

* Added more tests

* No revisions tests passing

* make BinariesAnalyzer the one responsible for the IDs

* remove build mode check

* Tests passing, only about 10 to complete

* changing how binary IDs are computed

* fixing tests

* refactoring tests

* binary ID belongs to graph Node

* working in graph

* Finished tests

* Upload package id not reference

* One more test

* SVN test

* IOError

* Fixing tests

* Fixing windows test with export package and package revisions

* Fixed test

* Pending migration

* Migration OK

* working...

* renaming conan->recipe

* Revisioning the revisions pull request

* CONAN_REVISIONS_ENABLED

* WIP

* Safer migration

* No revisions test

* V2 WIP

* WIP

* Review

* working except full_requires test broken

* WIP

* Rest routes

* Refactor routes

* Passing wihout revisions

* Removed time from metadata

* working but ConanInfo.full_requires test

* Passing tests

* Fixed more tests

* binary analyzer refactor

* Self review

* Imports

* small optimization of set_dirty(package_folder)

* Fixed rest api test

* Fixed slows

* Fixed py2 issue with exception translation with http codes

* rest api prepared revisions

* Dani and Diego review

* Fixing tests

* Review

* Imports

* renamed bid->package_id

* import and comment

* renamed test
