+++ conans/client/graph/graph_binaries.py
@@ -4,7 +4,8 @@ from conans.client.graph.graph import (BINARY_BUILD, BINARY_CACHE, BINARY_DOWNLO
                                        BINARY_SKIP, BINARY_UPDATE, BINARY_WORKSPACE,
                                        RECIPE_EDITABLE, BINARY_EDITABLE,
                                        RECIPE_CONSUMER, RECIPE_VIRTUAL)
-from conans.errors import NoRemoteAvailable, NotFoundException
+from conans.errors import NoRemoteAvailable, NotFoundException,\
+    conanfile_exception_formatter
 from conans.model.info import ConanInfo
 from conans.model.manifest import FileTreeManifest
 from conans.model.ref import PackageReference
@@ -30,12 +31,12 @@ class GraphBinariesAnalyzer(object):
                 output.warn("Current package is newer than remote upstream one")
 
     def _evaluate_node(self, node, build_mode, update, evaluated_nodes, remote_name):
-        assert node.binary is None, "Node binary is None"
+        assert node.binary is None, "Node.binary should be None"
+        assert node.package_id is not None, "Node.package_id shouldn't be None"
 
         ref, conanfile = node.ref, node.conanfile
-        package_id = conanfile.info.package_id()
-        pref = PackageReference(ref, package_id)
-        node.pref = pref
+        pref = PackageReference(ref, node.package_id)
+
         # Check that this same reference hasn't already been checked
         previous_node = evaluated_nodes.get(pref)
         if previous_node:
@@ -99,7 +100,7 @@ class GraphBinariesAnalyzer(object):
                     else:
                         if self._check_update(upstream_manifest, package_folder, output, node):
                             node.binary = BINARY_UPDATE
-                            node.pref = pref  # With revision
+                            node.prev = pref.revision  # With revision
                             if build_mode.outdated:
                                 info, pref = self._remote_manager.get_package_info(pref, remote)
                                 package_hash = info.recipe_hash()
@@ -135,7 +136,7 @@ class GraphBinariesAnalyzer(object):
 
             if remote_info:
                 node.binary = BINARY_DOWNLOAD
-                node.pref = pref  # With PREF
+                node.prev = pref.revision
                 package_hash = remote_info.recipe_hash
             else:
                 if build_mode.allowed(conanfile):
@@ -154,22 +155,55 @@ class GraphBinariesAnalyzer(object):
 
         node.binary_remote = remote
 
+    @staticmethod
+    def _compute_package_id(node):
+        conanfile = node.conanfile
+        neighbors = node.neighbors()
+        direct_reqs = []  # of PackageReference
+        indirect_reqs = set()   # of PackageReference, avoid duplicates
+        for neighbor in neighbors:
+            ref, nconan = neighbor.ref, neighbor.conanfile
+            pref = PackageReference(ref, neighbor.package_id)
+            direct_reqs.append(pref)
+            indirect_reqs.update(nconan.info.requires.refs())
+            conanfile.options.propagate_downstream(ref, nconan.info.full_options)
+            # Might be never used, but update original requirement, just in case
+            conanfile.requires[ref.name].ref = ref
+
+        # Make sure not duplicated
+        indirect_reqs.difference_update(direct_reqs)
+        # There might be options that are not upstream, backup them, might be
+        # for build-requires
+        conanfile.build_requires_options = conanfile.options.values
+        conanfile.options.clear_unused(indirect_reqs.union(direct_reqs))
+
+        conanfile.info = ConanInfo.create(conanfile.settings.values,
+                                          conanfile.options.values,
+                                          direct_reqs,
+                                          indirect_reqs)
+
+        # Once we are done, call package_id() to narrow and change possible values
+        with conanfile_exception_formatter(str(conanfile), "package_id"):
+            conanfile.package_id()
+
+        info = node.conanfile.info
+        node.package_id = info.package_id()
+
+    def _handle_private(self, node, deps_graph):
+        private_neighbours = node.private_neighbors()
+        if private_neighbours:
+            if node.binary in (BINARY_CACHE, BINARY_DOWNLOAD, BINARY_UPDATE):
+                for neigh in private_neighbours:
+                    neigh.binary = BINARY_SKIP
+                    closure = deps_graph.full_closure(neigh, private=True)
+                    for n in closure:
+                        n.binary = BINARY_SKIP
+
     def evaluate_graph(self, deps_graph, build_mode, update, remote_name):
         evaluated_nodes = {}
-        for node in deps_graph.nodes:
-            if node.recipe in (RECIPE_CONSUMER, RECIPE_VIRTUAL) or node.binary:
-                continue
-            private_neighbours = node.private_neighbors()
-            if private_neighbours:
-                self._evaluate_node(node, build_mode, update, evaluated_nodes, remote_name)
-                if node.binary in (BINARY_CACHE, BINARY_DOWNLOAD, BINARY_UPDATE):
-                    for neigh in private_neighbours:
-                        neigh.binary = BINARY_SKIP
-                        closure = deps_graph.full_closure(neigh, private=True)
-                        for n in closure:
-                            n.binary = BINARY_SKIP
-
-        for node in deps_graph.nodes:
-            if node.recipe in (RECIPE_CONSUMER, RECIPE_VIRTUAL) or node.binary:
+        for node in deps_graph.ordered_iterate():
+            self._compute_package_id(node)
+            if node.recipe in (RECIPE_CONSUMER, RECIPE_VIRTUAL):
                 continue
             self._evaluate_node(node, build_mode, update, evaluated_nodes, remote_name)
+            self._handle_private(node, deps_graph)
