+++ conans/client/installer.py
@@ -131,7 +131,7 @@ class _ConanPackageBuilder(object):
             source_folder = self.build_folder
         with get_env_context_manager(self._conan_file):
             install_folder = self.build_folder  # While installing, the infos goes to build folder
-            pkg_id = self._conan_file.info.package_id()
+            pkg_id = self._pref.id
             conanfile_path = self._cache.conanfile(self._ref)
 
             create_package(self._conan_file, pkg_id, source_folder, self.build_folder,
@@ -180,14 +180,14 @@ class _ConanPackageBuilder(object):
             with conanfile_exception_formatter(str(self._conan_file), "build"):
                 self._conan_file.build()
 
-            self._out.success("Package '%s' built" % self._conan_file.info.package_id())
+            self._out.success("Package '%s' built" % self._pref.id)
             self._out.info("Build folder %s" % self.build_folder)
             self._hook_manager.execute("post_build", conanfile=self._conan_file,
                                        reference=self._ref,
                                        package_id=self._pref.id)
         except Exception as exc:
             self._out.writeln("")
-            self._out.error("Package '%s' build failed" % self._conan_file.info.package_id())
+            self._out.error("Package '%s' build failed" % self._pref.id)
             self._out.warn("Build folder %s" % self.build_folder)
             if isinstance(exc, ConanExceptionInUserConanfileMethod):
                 raise exc
@@ -280,7 +280,7 @@ class BinaryInstaller(object):
             for node in level:
                 ref, conan_file = node.ref, node.conanfile
                 output = conan_file.output
-                package_id = conan_file.info.package_id()
+                package_id = node.package_id
                 if node.binary == BINARY_MISSING:
                     dependencies = [str(dep.dst) for dep in node.dependencies]
                     raise_package_not_found_error(conan_file, ref, package_id, dependencies,
@@ -336,16 +336,19 @@ class BinaryInstaller(object):
         with self._cache.package_lock(pref):
             if pref not in processed_package_references:
                 processed_package_references.add(pref)
-                set_dirty(package_folder)
                 if node.binary == BINARY_BUILD:
+                    set_dirty(package_folder)
                     self._build_package(node, pref, output, keep_build)
+                    clean_dirty(package_folder)
                 elif node.binary in (BINARY_UPDATE, BINARY_DOWNLOAD):
                     if not self._node_concurrently_installed(node, package_folder):
+                        set_dirty(package_folder)
                         assert pref.revision is not None, "Installer should receive #PREV always"
                         self._remote_manager.get_package(pref, package_folder,
                                                          node.binary_remote, output,
                                                          self._recorder)
                         self._registry.prefs.set(pref, node.binary_remote.name)
+                        clean_dirty(package_folder)
                     else:
                         output.success('Download skipped. Probable concurrent download')
                         log_package_got_from_local_cache(pref)
@@ -354,7 +357,7 @@ class BinaryInstaller(object):
                     output.success('Already installed!')
                     log_package_got_from_local_cache(pref)
                     self._recorder.package_fetched_from_cache(pref)
-                clean_dirty(package_folder)
+
             # Call the info method
             self._call_package_info(conan_file, package_folder)
             self._recorder.package_cpp_info(pref, conan_file.cpp_info)
