+++ conans/client/manifest_manager.py
@@ -20,13 +20,11 @@ class ManifestManager(object):
         if verify and not os.path.exists(self._paths.store):
             raise ConanException("Manifest folder does not exist: %s"
                                  % self._paths.store)
-        levels = graph.by_levels()
-        for level in levels:
-            for node in level:
-                if node.recipe in (RECIPE_CONSUMER, RECIPE_VIRTUAL):
-                    continue
-                self._handle_recipe(node, verify, interactive)
-                self._handle_package(node, verify, interactive)
+        for node in graph.ordered_iterate():
+            if node.recipe in (RECIPE_CONSUMER, RECIPE_VIRTUAL):
+                continue
+            self._handle_recipe(node, verify, interactive)
+            self._handle_package(node, verify, interactive)
 
     def _handle_recipe(self, node, verify, interactive):
         ref = node.ref
@@ -40,7 +38,7 @@ class ManifestManager(object):
 
     def _handle_package(self, node, verify, interactive):
         ref = node.ref
-        pref = PackageReference(ref, node.conanfile.info.package_id())
+        pref = PackageReference(ref, node.package_id)
         package_folder = self._cache.package(pref)
         read_manifest = FileTreeManifest.load(package_folder)
         expected_manifest = FileTreeManifest.create(package_folder)
