+++ conans/model/info.py
@@ -193,21 +193,17 @@ class RequirementsInfo(object):
             r.full_package_mode()
 
 
-class RequirementsList(list):
+class _PackageReferenceList(list):
     @staticmethod
     def loads(text):
-        return RequirementsList.deserialize(text.splitlines())
+        return _PackageReferenceList([PackageReference.loads(package_reference)
+                                     for package_reference in text.splitlines()])
 
     def dumps(self):
         return "\n".join(self.serialize())
 
     def serialize(self):
         return [str(r) for r in sorted(self)]
-
-    @staticmethod
-    def deserialize(data):
-        return RequirementsList([PackageReference.loads(package_reference)
-                                 for package_reference in data])
 
 
 class ConanInfo(object):
@@ -229,7 +225,7 @@ class ConanInfo(object):
         result.full_options = options
         result.options = options.copy()
         result.options.clear_indirect()
-        result.full_requires = RequirementsList(requires)
+        result.full_requires = _PackageReferenceList(requires)
         result.requires = RequirementsInfo(requires)
         result.requires.add(indirect_requires)
         result.full_requires.extend(indirect_requires)
@@ -251,7 +247,7 @@ class ConanInfo(object):
         result.full_settings = Values.loads(parser.full_settings)
         result.options = OptionsValues.loads(parser.options)
         result.full_options = OptionsValues.loads(parser.full_options)
-        result.full_requires = RequirementsList.loads(parser.full_requires)
+        result.full_requires = _PackageReferenceList.loads(parser.full_requires)
         result.requires = RequirementsInfo(result.full_requires)
         result.recipe_hash = parser.recipe_hash or None
 
@@ -312,18 +308,19 @@ class ConanInfo(object):
         """ The package_id of a conans is the sha1 of its specific requirements,
         options and settings
         """
-        computed_id = getattr(self, "_package_id", None)
-        if computed_id:
-            return computed_id
+        package_id = getattr(self, "_package_id", None)
+        if package_id:
+            return package_id
+
         result = []
         result.append(self.settings.sha)
         # Only are valid requires for OPtions those Non-Dev who are still in requires
-
         self.options.filter_used(self.requires.pkg_names)
         result.append(self.options.sha)
         result.append(self.requires.sha)
-        self._package_id = sha1('\n'.join(result).encode())
-        return self._package_id
+        package_id = sha1('\n'.join(result).encode())
+        self._package_id = package_id
+        return package_id
 
     def serialize_min(self):
         """
